+++ packages/actions/src/actionTypes/config.ts
@@ -14,11 +14,11 @@ export const DONE_SAVING_CONFIG     = "DONE_SAVING_CONFIG";
 export const MERGE_CONFIG           = "MERGE_CONFIG";
 export const SET_CONFIG_AT_KEY      = "SET_CONFIG_AT_KEY";
 
-export type LoadConfigAction        = Action<typeof LOAD_CONFIG>;                                      // FIXME: Name inconsistent
-export type SaveConfigAction        = Action<typeof SAVE_CONFIG>;                                      // FIXME: Name inconsistent
-export type DoneSavingConfigAction  = Action<typeof DONE_SAVING_CONFIG>;                               // FIXME: Name inconsistent
-export type MergeConfigAction       = Action<typeof MERGE_CONFIG,         ConfigPayload>;              // FIXME: Name inconsistent
-export type SetConfigAction     <T> = Action<typeof SET_CONFIG_AT_KEY,    { key: string; value: T }>;  // FIXME: Name inconsistent
+export type LoadConfigAction        = Action<typeof LOAD_CONFIG>;
+export type SaveConfigAction        = Action<typeof SAVE_CONFIG>;
+export type DoneSavingConfigAction  = Action<typeof DONE_SAVING_CONFIG>;
+export type MergeConfigAction       = Action<typeof MERGE_CONFIG,         ConfigPayload>;
+export type SetConfigAction     <T> = Action<typeof SET_CONFIG_AT_KEY,    { key: string; value: T }>;
 
 export const loadConfig             = makeActionFunction<LoadConfigAction>      (LOAD_CONFIG);
 export const saveConfig             = makeActionFunction<SaveConfigAction>      (SAVE_CONFIG);
