+++ packages/actions/src/actionTypes/host.ts
@@ -8,7 +8,7 @@ export const SET_APP_HOST     = "SET_APP_HOST";
 
 export type AddHost           = Action<typeof ADD_HOST,       HasHost &  { host: HostRecord }>;
 export type RemoveHost        = Action<typeof REMOVE_HOST,    HasHost>;
-export type SetAppHostAction  = Action<typeof SET_APP_HOST,   { host: HostRecord }>;  // FIXME: Name inconsistent
+export type SetAppHostAction  = Action<typeof SET_APP_HOST,   { host: HostRecord }>;
 
 export const addHost          = makeActionFunction<AddHost>         (ADD_HOST);
 export const removeHost       = makeActionFunction<RemoveHost>      (REMOVE_HOST);
